// third_party/node-banana/upstream/src/types/nodes.ts
export interface ModelInputDef {
  name: string;
  type: "image" | "text" | "audio" | "video";
  required: boolean;
  label: string;
  description?: string;
}